パッケージ名、変数名の変更
