--- src/chapter02_task/Task01_HelloWorld.java
@@ -1,8 +1,0 @@
-package chapter02_task;
-
-public class Task01_HelloWorld {
-	public static void main(String[] args) {
-		String greeting = "Hello World";
-		System.out.println(greeting);
-	}
-}
