--- src/chapter02_task/Task02_sum.java
@@ -1,12 +1,0 @@
-package chapter02_task;
-
-public class Task02_sum {
-	public static void main(String[] args) {
-		int sum = 0;
-
-		for(int i =1; i<=10; i++) {
-			sum += i; 
-		}
-		System.out.println(sum);
-	}
-}
