--- src/chapter02_task/Task02_sumAPIver.java
@@ -1,8 +1,0 @@
-package chapter02_task;
-import java.util.stream.IntStream;
-
-public class Task02_sumAPIver {
-	public static void main(String[] args) {
-		System.out.println(IntStream.rangeClosed(1, 10).sum());
-	}
-}
