--- src/chapter02_task/Task04.java
@@ -1,16 +1,0 @@
-package chapter02_task;
-
-public class Task04 {
-	public static void main(String[] args) {
-		int n = 0;
-		if (n == 0) {
-			System.out.println("偽");
-		}
-		else if (n == 1) {
-			System.out.println("真");
-		}
-		else {
-			System.out.println("ERROR");
-		}
-	}
-}
