+++ src/chapter02task/Task01HelloWorld.java
@@ -1,6 +1,6 @@
-package chapter02task;
+package chapter02_task;
 
-public class Task01HelloWorld {
+public class Task01_HelloWorld {
 	public static void main(String[] args) {
 		String greeting = "Hello World";
 		System.out.println(greeting);
