+++ src/chapter02task/Task02sumAPIver.java
@@ -1,7 +1,7 @@
-package chapter02task;
+package chapter02_task;
 import java.util.stream.IntStream;
 
-public class Task02sumAPIver {
+public class Task02_sumAPIver {
 	public static void main(String[] args) {
 		System.out.println(IntStream.rangeClosed(1, 10).sum());
 	}
