+++ src/chapter02task/Task03threeYearsLater.java
@@ -1,8 +1,9 @@
-package chapter02task;
+package chapter02_task;
+
 import java.time.DayOfWeek;
 import java.time.LocalDateTime;
 
-public class Task03threeYearsLater {
+public class Task03_3yearsLater {
 	public static void main(String[] args) {
 		//現在日付
 		LocalDateTime currentDateTime = LocalDateTime.now();
